+++ src/main/java/com/example/backend_projedata/model/ProductCompositionResponseDTO.java
@@ -1,4 +1,4 @@
 package com.example.backend_projedata.model;
 
-public record ProductCompositionResponseDTO(String product,String rawMaterial,double quantity_required) {
+public record ProductCompositionResponseDTO(String product,String rawMaterial,Long quantity_required) {
 }
